pi/pi_requirements.txt
flask
opencv-python
qrcode
Pillow
picamera2
